Vibration adjusted and Sync Issue resolution
There was a bug because sometimes the 'stop' bit received from the console varies by about 10-20 microseconds. Increased the threshold to detect stop bits up to 2.5ish microseconds. Much more stable this way it seems. Also added in a desync resolution function which adds a delay in the case of a bad command received.