Software Snapback Update
This update is the first to play with the idea of software snapback. 

This update breaks compatibility with smashscope as the operations take too long for how often smashscope polls.

This operation runs perfect at a wide range of 1ms to 16ms polling intervals. This means it functions between GameCube as well as Nintendo Switch OEM USB adapter. 

A default hardware capacitance value total of around 0.85uf - 1uf is recommended alongside this software filter. 

Snapback can be adjusted in this demo build mid-play by holding minus and pressing ZL and ZR to decrease/increase the amount of snapback being applied. a range of 1-16 values, with 1 having NO filter, and 16 only being the filter.